requirements.txt
fastapi>=0.135.3
uvicorn>=0.44.0
sqlalchemy>=2.0.49
pydantic>=2.12.5
pydantic-settings>=2.13.1
argon2-cffi>=25.1.0
PyJWT>=2.12.1
pydantic[email]
python-multipart>=0.0.24